refactor(MainController): Application에서 컨트롤 로직 분리.
- 메인 화면에서 이루어지는 로직에 대한 컨트롤러의 역할을 하고 있던 Application의 코드를 MainController를 만들어 분리했습니다.

- 프로그램이 전반적으로 Main/Station/Line/Section이라는 영역으로 나뉘어 돌아가는 만큼, MainController도 분리하는게 맞다고 생각했습니다.

- 그리고 MainController는 스스로 매칭되는 MainView가 있고, View 간에는 계층화가 되어있기 때문에 직접 OutputView를 호출하지 않도록 코드를 일부 수정했습니다.
